Upgrade seeds to latest commands
