+++ api/package.json
@@ -6,23 +6,28 @@
   "dependencies": {
     "babel-core": "^6.22.1",
     "body-parser": "^1.15.0",
+    "casual": "^1.5.12",
     "cors": "^2.7.1",
     "eventstore-node": "^0.0.25",
     "express": "^4.13.4",
     "glob": "^7.0.5",
     "on-response": "^1.0.1",
+    "require-directory": "^2.1.1",
     "uuid": "latest",
     "validator": "^6.2.0"
   },
   "devDependencies": {
     "assert": "^1.4.1",
     "babel-cli": "^6.11.4",
-    "babel-plugin-transform-object-rest-spread": "^6.20.2",
-    "babel-preset-es2015": "^6.13.2",
-    "babel-preset-es2015-node": "^6.1.1",
-    "babel-preset-react": "^6.3.13",
+    "babel-plugin-syntax-async-functions": "^6.13.0",
+    "babel-plugin-transform-async-to-generator": "^6.22.0",
+    "babel-plugin-transform-es2015-modules-commonjs": "^6.23.0",
+    "babel-plugin-transform-object-rest-spread": "^6.23.0",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "babel-preset-es2015": "^6.24.1",
     "concurrently": "^3.1.0",
-    "mocha": "^3.2.0"
+    "mocha": "^3.2.0",
+    "newman": "^3.5.2"
   },
   "scripts": {
     "clean": "rm -rf dist",
@@ -35,25 +40,14 @@
     "test-api": "newman run noneoftheabove.postman_collection.json -e noneoftheabove-localhost-postman-environment.postman_environment.json",
     "test-debug": "mocha --debug-brk --compilers js:babel-core/register",
     "start": "npm run server",
-    "server": "babel-node --presets es2015 src/app local"
+    "server": "babel-node src/app local",
+    "seed": "babel-node scripts/seed.js"
   },
   "keywords": [
     "voting",
     "blockchain",
     "democracy"
   ],
-  "babel": {
-    "presets": [
-      "es2015",
-      "react"
-    ],
-    "plugins": [
-      "transform-object-rest-spread"
-    ],
-    "sourceMaps": true,
-    "retainLines": true,
-    "ignore": "**/TODO_*.js"
-  },
   "author": "Adaptech Solutions",
   "license": "UNLICENSED",
   "private": true
